refactor(tanstack-query): tighten TransactionOperation typing and split react tests
- Make `args` required on TransactionOperation only for ops whose args type has required fields; keep it optional for read ops with all-optional args
- Skip read-only ops in transaction onSuccess so they don't reach createInvalidator
- Split the 2k-line react-query test file into per-feature files under test/react/, adding shared helpers
- Add typing tests for args optionality and delegate-model rejection in the sequential transaction suite
